Translate all sensible builtins
